+++ app/(tabs)/progress.tsx
@@ -1,12 +1,15 @@
 import { ThemedText } from '@/components/ThemedText';
 import { ThemedView } from '@/components/ThemedView';
-import { StyleSheet } from 'react-native';
+import { StyleSheet, Platform } from 'react-native';
+import { LinearGradient } from 'expo-linear-gradient';
+import { AppColors } from '@/constants/Colors';
+import ScreenHeader from '@/components/ScreenHeader';
 
 export default function ProgressScreen() {
     return (
-        <ThemedView style={styles.title}>
-            <ThemedText>Progress</ThemedText>
-        </ThemedView>
+        <LinearGradient style={{ flex: 1, paddingTop: Platform.OS == 'ios' ? 50 : 0}} colors={[AppColors.OffWhite, AppColors.LightBlue]}>
+        <ScreenHeader title="Progress" />
+      </LinearGradient>
     )
 }