+++ constants/Colors.ts
@@ -2,16 +2,29 @@
  * Below are the colors that are used in the app. The colors are defined in the light and dark mode.
  * There are many other ways to style your app. For example, [Nativewind](https://www.nativewind.dev/), [Tamagui](https://tamagui.dev/), [unistyles](https://reactnativeunistyles.vercel.app), etc.
  */
-
 const tintColorLight = '#0a7ea4';
 const tintColorDark = '#fff';
+
+export const AppColors = {
+  Red: '#E91313',
+  Green: '#4AB25B',
+  Blue: '#3764BE',
+  Purple: '#B5A6E7',
+  LightBlue: '#ACD3E0',
+  DarkBlue: '#2C5999',
+  BlueGray: '#5A7AA5',
+  Pink: '#EB9BD0',
+  Black: '#000000',
+  White: '#ffffff',
+  OffWhite: '#F1F6F2',
+};
 
 export const Colors = {
   light: {
     text: '#11181C',
-    background: '#fff',
+    background: "#fff",
     tint: tintColorLight,
-    icon: '#000000',
+    icon: '#687076',
     tabIconDefault: '#000000',
     tabIconSelected: tintColorLight,
   },