+++ CHANGELOG.md
@@ -8,3 +8,16 @@
 - Readme
 - Node dependencies
 
+## Feb 28
+### Added
+- cleck.js sdk
+- Authentication Signup/SignIn setup
+- custom font(Meticula)
+
+## March 1
+### Added
+- Logout button to profile page
+
+### Updated
+- Signup/Signin page to include error handling
+- Privacy policy link in signup/signin pages